config struct

This is the object holding a complete collection of all parsed configuration parameters and their defaults, if not specified.

MOLGA.Configuration.ConfigurationObject — Type

Stores the entire configuration parameters.

initial_population: See InitialPopulationConfiguration.

charge: Charge of the structure.

spin_multiplicity: Spin multiplicity of the structure.

distance_thresholds: See DistanceThresholds.

preserve_fragments: True if fragments should be preserved during recombination and mutation.

mutation: See MutationConfiguration.

MOLGA.Configuration.InitialPopulationConfiguration — Type

Defines the configuration of the initial population.

num_structures::Int64: The number of initial structures to be created.

box_size::StaticArraysCore.SVector{3, Float64}: The three dimensions of the periodic boundary condition box (in angstroms).

atom_config::Array{Union{MOLGA.Configuration.ConfigAtom, MOLGA.Configuration.ConfigMolecule}}: See AtomConfig.