API Reference

This page provides a list of all documented functions, structs and modules in the MolecularGeneticAlgorithm. Note that this is list is auto-generated using Documenter's @autodocs function.

MOLGA.Configuration.create_atom_config

MOLGA.Configuration.load_config

MOLGA.GeneticAlgorithm.InitialPopulation.create

MOLGA.GeneticAlgorithm.InitialPopulation.find_position!

MOLGA.GeneticAlgorithm.InitialPopulation.find_position!

MOLGA.GeneticAlgorithm.InitialPopulation.generate_atoms

MOLGA.GeneticAlgorithm.InitialPopulation.generate_indices

MOLGA.GeneticAlgorithm.Mutation.mutate!

MOLGA.GeneticAlgorithm.Mutation.reposition_atom!

MOLGA.GeneticAlgorithm.Mutation.reposition_atom_preserving!

MOLGA.GeneticAlgorithm.Mutation.switch_atom!

MOLGA.GeneticAlgorithm.Mutation.switch_atom_by_indices!

MOLGA.GeneticAlgorithm.Mutation.switch_atom_preserving!

MOLGA.GeneticAlgorithm.Mutation.try_reposition_atom_preserving!

MOLGA.GeneticAlgorithm.Mutation.try_switch_atom_preserving!

MOLGA.GeneticAlgorithm.Utils.check_distance

MOLGA.GeneticAlgorithm.Utils.random_angles

MOLGA.GeneticAlgorithm.Utils.random_position

MOLGA.GeneticAlgorithm.Utils.rotate_position

MOLGA.GeneticAlgorithm.Utils.shuffle_rows!

MOLGA.Helpers.atomic_number

MOLGA.Helpers.element_symbol

MOLGA.IOm.export_xyz

MOLGA.IOm.load_xyz

MOLGA.run_genetic_algorithm

MOLGA.Configuration.AtomConfig

MOLGA.Configuration.ConfigAtom

MOLGA.Configuration.ConfigMolecule

MOLGA.Configuration.ConfigurationObject

MOLGA.Configuration.DistanceThresholds

MOLGA.Configuration.InitialPopulationConfiguration

MOLGA.Configuration.MutationConfiguration

MOLGA.Types.AbstractAtom

MOLGA.Types.Atom

MOLGA.Types.BaseAtom

MOLGA.Types.Structure

MOLGA.Types.Vec

Main Module

MOLGA.run_genetic_algorithm — Function

run_genetic_algorithm(configuration_object [; debug_log = false, rng])

Run the molecular genetic algorithm using the provided ConfigurationObject.

Parameters

configuration_object::ConfigurationObject

debug_log::Bool: Defines if you want to enable the logging of debug-level events.

rng::AbstractRNG: Random number generator. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

rungeneticalgorithm(configurationfile [, debuglog = false, rng])

Start the molecular genetic algorithm and load the necessary configuration parameters from the specified TOML file.

Parameters

configuration_file::String: The filename and path to the TOML configuration file (see specification here).

debug_log::Bool: Defines if you want to enable the logging of debug-level events.

rng::AbstractRNG: Random number generator. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

Genetic Algorithm

MOLGA.GeneticAlgorithm.InitialPopulation.create — Method

create(initial_population_config, distance_thresholds [; rng])

Initialize the population by creating the desired amount of random structures using the InitialPopulationConfiguration.

Arguments

initial_population_config::InitialPopulationConfiguration

distance_thresholds::DistanceThresholds

rng::AbstractRNG: Random number generator. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

Returns

A vector of structures.

MOLGA.GeneticAlgorithm.InitialPopulation.find_position! — Method

find_position!(atom, atom_list, permanent_index, fragment_kind box_size, distance_thresholds,
    max_attempts, [; rng])

Try to find a position for an atom within the periodic boundary condition box by respecting distance thresholds. If a proper position is found, the list of atoms is updated in-place by adding the specified atom with its permanent_index and fragment_kind to the atom_list.

Arguments

atom::ConfigAtom: The atom to be placed into the structure.

atom_list::Vector{Atom}: The list of atoms of the structure.

permanent_index::Int

fragment_kind::Int

box_size::Vec

distance_thresholds::DistanceThresholds

max_attempts::Int: Maximum number of attempts to find a proper random position.

rng::AbstractRNG: Random number generator. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

Returns

true if the atom could be placed,

false if no suitable position could be found within the maximum number of attempts.

Example

julia> atom_list = [
           Atom(8, [-0.299181461, 0.000000000, -0.473896792], 1, 1),
           Atom(1, [-0.299181461, 0.759337000, 0.122146208], 1, 1),
           Atom(1, [-0.299181461, -0.759337000, 0.122146208], 1, 1),
       ];

julia> find_position!(
           ConfigAtom(1, 1),
           atom_list,
           2,
           2,
           Vec([8, 8, 8]),
           DistanceThresholds(1, 5),
           5_000;
           rng=Xoshiro(123),
       )
true

julia> atom_list
4-element Vector{Atom}:
 O     1    1 [-0.299181461         0.0                  -0.473896792        ]

 H     1    1 [-0.299181461         0.759337             0.122146208         ]

 H     1    1 [-0.299181461         -0.759337            0.122146208         ]

 H     2    2 [0.1697103642830644   0.6944540596267874   3.1270295847422487  ]

MOLGA.GeneticAlgorithm.InitialPopulation.find_position! — Method

find_position!(fragment, atom_list, permanent_index, fragment_kind, box_size,
    distance_thresholds, max_attempts [; rng])

Place a fragment (ConfigMolecule) into the periodic boundary condition box. Perform distance checks for all atoms of the fragment and assign the specified permanent index and fragment kind to all atoms. This method upates the atom_list in-place.

See find_position!(atom, ...) for more details.

MOLGA.GeneticAlgorithm.InitialPopulation.generate_atoms — Method

generate_atoms(indices, population_config, distance_thresholds, max_iterations [; rng])

Distribute the atoms and fragments randomly within a cuboid periodic boundary condition box with respect to the distance thresholds.

Arguments

indices::Matrix{Int}: A two-dimensional matrix generated by generate_indices representing the atom configuration.

population_config::InitialPopulationConfiguration

distance_thresholds::DistanceThresholds

max_iterations::Int: The maximum number of iterations to find proper positions while distributing the atoms and fragments.

rng::AbstractRNG: Random number generator. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

Returns

A vector of atoms.

Example

Small example creating an atom list with one Ag atom and four water molecules.

julia> atom_config::AtomConfig = [
           ConfigAtom(1, 47), # 1x Ag
           ConfigMolecule( # 4x H2O
               4,
               [
                   BaseAtom(8, [-1.674872668 0.000000000 -0.984966492]),
                   BaseAtom(1, [-1.674872668 0.759337000 -0.388923492]),
                   BaseAtom(1, [-1.674872668 -0.759337000 -0.388923492]),
               ],
           ),
       ];

julia> indices = [1 1; 2 2; 2 3; 2 4; 2 5];
       population_config = InitialPopulationConfiguration(2, Vec(8, 8, 8), atom_config);
       thresholds = DistanceThresholds(0.6, 5);

julia> atoms = generate_atoms(indices, population_config, thresholds, 100; rng=Xoshiro(123));

julia> show(atoms)
   frag perm position
O     2    5 [-2.4770333452109243  -3.431281259074869   4.54646256001069    ]
H     2    5 [-1.7648962458876567  -2.9287153499858496  4.961370015001522   ]
H     2    5 [-2.1826188648443647  -4.3478380096041835  4.617903264287958   ]
O     2    4 [0.7497562286266901   -3.2644362986611446  2.294185463046077   ]
H     2    4 [1.0913043037934056   -2.365109811952456   2.214081030758482   ]
H     2    4 [1.5403392220307977   -3.7973137270056903  2.445447204141124   ]
O     2    2 [0.009263151100168399 -2.841697310728465   5.348068954930102   ]
H     2    2 [0.16930600494596648  -3.7718340771241343  5.145357785652479   ]
H     2    2 [-0.6139073734799423  -2.5622881483814584  4.665830759935792   ]
O     2    3 [0.09878867117371559  -2.7223525995269084  1.934017529032187   ]
H     2    3 [-0.45581965769155497 -3.094652410991112   2.630912348466935   ]
H     2    3 [-0.08138337062727174 -1.775059413996567   1.9791129103815164  ]
Ag    1    1 [1.436302063959932    -3.3577497929871196  1.9092560648370087  ]

MOLGA.GeneticAlgorithm.InitialPopulation.generate_indices — Method

generate_indices(atom_config)

Create a list of the indices using the AtomConfig.

The resulting two-dimensional matrix contains the fragment_kind in the first column and the permanent_index in the second column. Both indices are integers starting from $1$. The entries are inserted as often as specified in the AtomConfig quantity-property.

Each fragment or single atom (one or more) have their own fragment_kind which is equivalent to the index of the AtomConfig list. To provide an option to preserve fragments, all atoms of an instance of a fragment have a common permanent_index.

Example

Small example distributing two hydrogen atoms, one oxygen atom, as well as two carbon dioxide fragments for demonstration purposes.

atom_config::AtomConfig = [
    ConfigAtom(2, 1), # 2x H
    ConfigAtom(1, 8), # 1x O
    ConfigMolecule(
        2, # 2x CO2
        [
            BaseAtom(6, [-1.161937300 0.000000000 -1.704638554]),
            BaseAtom(8, [-0.610097929 -0.126832008 -0.681728275]),
            BaseAtom(8, [-1.713776672 0.126832008 -2.727548833]),
        ],
    ),
]
generate_indices(atom_config)

# output

5×2 Matrix{Int64}:
 1  1
 1  2
 2  3
 3  4
 3  5

MOLGA.GeneticAlgorithm.Mutation.mutate! — Method

mutate!(population, mutation_config, preserve_fragments, box_size, distance_thresholds
    [; mutation_method_distribution, rng])

Perform mutation on all structures of the population in-place.

Currently, there are two mutations (atom switch, atom reposition). Each mutation function is applied individually to each structure with the specified mutation probability, i.e. it is possible for a structure to be mutated twice (first atom switch, then atom reposition).

If the keep_non_mutated switch is activated, the original structure is added to the population before performing the mutation(s).

Depending on the choice of the preserve_fragments option, the appropriate function is selected for both mutations.

Arguments

population::Vector{Structure}

mutation_config::MutationConfiguration

preserve_fragments::Bool: If you want to preserve fragments.

box_size::Vec

distance_thresholds::DistanceThresholds

mutation_method_distribution::AbstractVector{Real}: Define the distribution of the two mutation methods. The first position controls the atom switch mutation, the second the atom reposition. Defaults to [1, 1] which means that both mutations have the same probability.

rng::AbstractRNG: Random number generator. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

MOLGA.GeneticAlgorithm.Mutation.reposition_atom! — Method

reposition_atom!(structure, box_size [; rng])

Move a randomly chosen atom from the provided structure in-place to another random position within the cuboid box. This doesn't preserve any fragments.

Arguments

structure::Structure

box_size::Vec

rng::AbstractRNG: Random number generator. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

Example

julia> atoms = [
           Atom(8, [-0.702728547, 0.000000000, -3.588916226], 1, 1),
           Atom(6, [-0.702735805, -0.000007025, -2.382267226], 2, 2),
           Atom(1, [-0.702724918, 0.937663512, -1.787406767], 3, 3),
           Atom(1, [-0.702724918, -0.937656488, -1.787385685], 4, 3),
       ];

julia> structure = Structure(; atoms)
Structure with 4 atoms
Energy: 0.0
Nuclear repulsion energy: 0.0
Age: 0
   frag perm position
O     1    1 [-0.702728547         0.0                  -3.588916226        ]
C     2    2 [-0.702735805         -7.025e-6            -2.382267226        ]
H     3    3 [-0.702724918         0.937663512          -1.787406767        ]
H     3    4 [-0.702724918         -0.937656488         -1.787385685        ]

julia> reposition_atom!(structure, Vec(8, 8, 8); rng=Xoshiro(123))

julia> structure
Structure with 4 atoms
Energy: 0.0
Nuclear repulsion energy: 0.0
Age: 0
   frag perm position
O     1    1 [-0.702728547         0.0                  -3.588916226        ]
C     2    2 [-0.702735805         -7.025e-6            -2.382267226        ]
H     3    3 [0.6944540596267874   3.1270295847422487   -2.472746407793897  ]
H     3    4 [-0.702724918         -0.937656488         -1.787385685        ]

MOLGA.GeneticAlgorithm.Mutation.reposition_atom_preserving! — Method

reposition_atom_preserving!(structure, box_size, distance_thresholds [; rng])

Move a randomly chosen atom or fragment from the provided structure in-place to another random position and preserve any fragments.

All atoms are tried out in a randomized order. If there is no successful reposition, the structure is not modified.

Arguments

structure::Structure

box_size::Vec

distance_thresholds::DistanceThresholds

rng::AbstractRNG: Random number generator. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

MOLGA.GeneticAlgorithm.Mutation.switch_atom! — Method

switch_atom!(structure [; rng])

Exchange the position of two randomly selected atoms (of different chemical elements) from the provided structure in-place without preserving any fragments.

Arguments

structure::Structure

rng::AbstractRNG: Random number generator. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

Example

julia> atoms = [
           Atom(8, [-0.702728547, 0.000000000, -3.588916226], 1, 1),
           Atom(6, [-0.702735805, -0.000007025, -2.382267226], 2, 2),
           Atom(1, [-0.702724918, 0.937663512, -1.787406767], 3, 3),
           Atom(1, [-0.702724918, -0.937656488, -1.787385685], 4, 3),
       ];

julia> structure = Structure(; atoms)
Structure with 4 atoms
Energy: 0.0
Nuclear repulsion energy: 0.0
Age: 0
   frag perm position
O     1    1 [-0.702728547         0.0                  -3.588916226        ]
C     2    2 [-0.702735805         -7.025e-6            -2.382267226        ]
H     3    3 [-0.702724918         0.937663512          -1.787406767        ]
H     3    4 [-0.702724918         -0.937656488         -1.787385685        ]

julia> switch_atom!(structure; rng=Xoshiro(123))

julia> structure
Structure with 4 atoms
Energy: 0.0
Nuclear repulsion energy: 0.0
Age: 1
   frag perm position
O     1    1 [-0.702724918         0.937663512          -1.787406767        ]
C     2    2 [-0.702735805         -7.025e-6            -2.382267226        ]
H     3    3 [-0.702728547         0.0                  -3.588916226        ]
H     3    4 [-0.702724918         -0.937656488         -1.787385685        ]

MOLGA.GeneticAlgorithm.Mutation.switch_atom_by_indices! — Method

switch_atom_by_indices!(structure, index_1, index_2)

Exchange the position of the index-specified atoms from the provided structure in-place without preserving any fragments.

Arguments

structure::Structure

index_1::Int: Index of the first atom.

index_2::Int: Index of the second atom.

Example

julia> atoms = [
           Atom(8, [-0.702728547, 0.000000000, -3.588916226], 1, 1),
           Atom(6, [-0.702735805, -0.000007025, -2.382267226], 2, 2),
           Atom(1, [-0.702724918, 0.937663512, -1.787406767], 3, 3),
           Atom(1, [-0.702724918, -0.937656488, -1.787385685], 4, 3),
       ];

julia> structure = Structure(; atoms)
Structure with 4 atoms
Energy: 0.0
Nuclear repulsion energy: 0.0
Age: 0
   frag perm position
O     1    1 [-0.702728547         0.0                  -3.588916226        ]
C     2    2 [-0.702735805         -7.025e-6            -2.382267226        ]
H     3    3 [-0.702724918         0.937663512          -1.787406767        ]
H     3    4 [-0.702724918         -0.937656488         -1.787385685        ]

julia> switch_atom_by_indices!(structure, 1, 4)

julia> structure
Structure with 4 atoms
Energy: 0.0
Nuclear repulsion energy: 0.0
Age: 1
   frag perm position
O     1    1 [-0.702724918         -0.937656488         -1.787385685        ]
C     2    2 [-0.702735805         -7.025e-6            -2.382267226        ]
H     3    3 [-0.702724918         0.937663512          -1.787406767        ]
H     3    4 [-0.702728547         0.0                  -3.588916226        ]

MOLGA.GeneticAlgorithm.Mutation.switch_atom_preserving! — Method

switch_atom_preserving!(structure, distance_thresholds [; rng])

Exchange the position of two randomly selected atoms and/or fragments from the provided structure in-place while preserving fragments.

We loop over all possible combinations in a random order and try the switch. If it does't work, the structure is not modified.

Arguments

structure::Structure

distance_thresholds::DistanceThresholds

rng::AbstractRNG: Random number generator. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

MOLGA.GeneticAlgorithm.Mutation.try_reposition_atom_preserving! — Method

try_reposition_atom_preserving!(structure, atom_index, new_position, distance_thresholds)

Try to move the specified atom or fragment from the provided structure to the provided position, preserving fragments.

While building up the new atom list, distance checks are subsequently performed. Here, we only check the lower distance threshold.

Arguments

structure::Structure

atom_index::Int: Index of the atom to be moved.

new_position::Vec

distance_thresholds::DistanceThresholds

Returns

true if the reposition is successful,

false if a distance check fails.

MOLGA.GeneticAlgorithm.Mutation.try_switch_atom_preserving! — Method

try_switch_atom_preserving!(structure, index_1, index_2, distance_thresholds)

Try to exchange the position of the atoms and/or fragments, specified by their indices, in-place with preservation of fragments.

While building up the new atom list, distance checks are subsequently performed. Here, we only check the lower distance threshold.

Arguments

structure::Structure

index_1::Int: Index of the first atom.

index_2::Int: Index of the second atom.

distance_thresholds::DistanceThresholds

Returns

true if the switch is successful,

false if a distance check fails.

MOLGA.GeneticAlgorithm.Utils.check_distance — Function

check_distance(new_position, atoms, thresholds, check_only_lower_threshold=false)

Checks if the new position is sufficiently far away from all existing atoms in the atom list.

Arguments

new_position::Vec: The new position to be checked.

atoms::Vector{<:AbstractAtom}: The current list of atoms.

thresholds::DistanceThresholds

check_only_lower_threshold::Bool: True if only the lower distance threshold should be checked.

Example

The atom list contains the atoms of a water molecule. The task is to check if the chosen position is suitable. The distances between the new position and the hydrogen atoms are 1.80392 and 1.30274; the distance between pos and the carbon atom is 1.40762 (all measured with Chemcraft).

julia> pos = Vec([0.959329065, 0.512611795, -0.106788996]);

julia> check_distance(pos, Atom[], DistanceThresholds(0.5, 5.0))
true

julia> atoms = [
           BaseAtom(8, [-0.299181461, 0.000000000, -0.473896792]),
           BaseAtom(1, [-0.299181461, 0.759337000, 0.122146208]),
           BaseAtom(1, [-0.299181461, -0.759337000, 0.122146208]),
       ];

julia> check_distance(pos, atoms, DistanceThresholds(0.5, 5.0))
true

julia> check_distance(pos, atoms, DistanceThresholds(1.4, 5.0))
false

julia> check_distance(pos, atoms, DistanceThresholds(0.5, 1))
false

julia> check_distance(pos, atoms, DistanceThresholds(0.5, 1), true)
true

MOLGA.GeneticAlgorithm.Utils.random_angles — Method

random_angles([; rng])

Generate a vector of three random angles (from $0$ to $2\pi$, in radians).

You can pass a random number generator to rng. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

Example

julia> random_angles(; rng=Xoshiro(123))
3-element StaticArraysCore.SVector{3, Float64} with indices SOneTo(3):
 3.2748828620072237
 3.687015596584574
 5.597555946335841

MOLGA.GeneticAlgorithm.Utils.random_position — Method

random_position(box_size [; rng])

Generate a random position within the specified box. The origin is the box's center point, so that a box size of $\mathbf{s}=\left(s_x,s_y,s_z\right)$ leads to a random position $\mathbf{x}$

You can pass a random number generator to rng. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

\[-\frac{1}{2}\,\mathbf{s}\leq\mathbf{x}\leq\frac{1}{2}\,\mathbf{s} \text{.}\]

Example

julia> random_position(Vec([8, 5, 5]); rng=Xoshiro(1))
3-element StaticArraysCore.SVector{3, Float64} with indices SOneTo(3):
 -3.413069164245657
 -0.7537925522140694
  0.9941334184573423

MOLGA.GeneticAlgorithm.Utils.rotate_position — Method

rotate_position(position, angle)

Rotate a 3D position vector by a 3D angle vector around the origin using rotation matrices.

Arguments

position::Vector{Float64}: The 3D position vector to be rotated.

angle::Vector{Float64}: The 3D angle vector representing rotation angles around the $x$, $y$, and $z$ axes in radians.

Example

julia> rotate_position(Vec([1, 0, 0.5]), Vec([0, 0, π]))
3-element StaticArraysCore.SVector{3, Float64} with indices SOneTo(3):
 -1.0
  1.2246467991473532e-16
  0.5

MOLGA.GeneticAlgorithm.Utils.shuffle_rows! — Method

shuffle_rows!(m::AbstractMatrix [; rng])

You can pass a random number generator to rng. If you need consistent results for testing purposes, pass a seeded pseudorandom number generator here, eg. Xoshiro(seed). Defaults to Random.default_rng().

Example

julia> mat = [1 10; 2 20; 3 30; 4 40; 5 50];

julia> shuffle_rows!(mat; rng=Xoshiro(123));

julia> mat
5×2 Matrix{Int64}:
 5  50
 4  40
 2  20
 3  30
 1  10

Interfaces

Types

MOLGA.Types.AbstractAtom — Type

Abstract type for Atom and BaseAtom.

MOLGA.Types.Atom — Type

Atom(element, position, permanent_index, fragment_kind)

Stores properties of a single atom.

element::Int64: Atomic number of the element, eg. 6 for Carbon.

position::StaticArraysCore.SVector{3, Float64}: 3D position vector of the atom in angstroms.

permanent_index::Int64: Unique index attached to an atom.

fragment_kind::Int64: Kind of fragment this atom belongs to, e.g. all atoms that belong to a H2O molecule have the same value here, also all single O atoms have the same value here, respectively.

MOLGA.Types.BaseAtom — Type

BaseAtom(element, position)

Stores a basic atom.

element::Int64: Atomic number of the element, eg. 6 for Carbon.

position::StaticArraysCore.SVector{3, Float64}: 3D position vector of the atom in angstroms.

MOLGA.Types.Structure — Type

Structure(atoms)
Structure(atoms [, energy [, nuclear_repulsion_energy [, age]]])

Stores properties of a molecular structure.

atoms::Vector{MOLGA.Types.Atom}: Array of atoms that the structure consists of.

energy::Float64: Energy of the structure in eV.

nuclear_repulsion_energy::Float64: Nuclear repulsion energy of the structure in eV.

age::Int64: Number of cycles that this structure is already present.

MOLGA.Types.Vec — Type

Three-dimensional vector.

Configuration

MOLGA.Configuration.AtomConfig — Type

The atom configuration consists of single atoms and molecules.

MOLGA.Configuration.ConfigAtom — Type

ConfigAtom(num, element)

Atom to define initial atom configuration.

quantity::Int64: Number of atoms of this kind.

element::Int64: Chemical element denoted as atomic number.

MOLGA.Configuration.ConfigMolecule — Type

ConfigMolecule(num, atoms)

Molecule to define initial atom configuration.

quantity::Int64: Number of molecules of this kind.

atoms::Vector{MOLGA.Types.BaseAtom}: Array of atoms that the molecule consists of.

MOLGA.Configuration.ConfigurationObject — Type

Stores the entire configuration parameters.

initial_population: See InitialPopulationConfiguration.

charge: Charge of the structure.

spin_multiplicity: Spin multiplicity of the structure.

distance_thresholds: See DistanceThresholds.

preserve_fragments: True if fragments should be preserved during recombination and mutation.

mutation: See MutationConfiguration.

MOLGA.Configuration.DistanceThresholds — Type

DistanceThresholds(min, max)

Defines the thresholds for distances between atoms.

min::Float64: Minimum distance that two atoms need.

max::Float64: Maximum distance that two atoms need.

The values are taken into account at several points, including in the generation of the initial population and in mutation and recombination. They are given in angstroms.

MOLGA.Configuration.InitialPopulationConfiguration — Type

Defines the configuration of the initial population.

num_structures::Int64: The number of initial structures to be created.

box_size::StaticArraysCore.SVector{3, Float64}: The three dimensions of the periodic boundary condition box (in angstroms).

atom_config::Array{Union{MOLGA.Configuration.ConfigAtom, MOLGA.Configuration.ConfigMolecule}}: See AtomConfig.

MOLGA.Configuration.MutationConfiguration — Type

Configure the mutation parameters.

probability::Float64: Probability in the interval [0,1] with which each individual structure mutates. Each mutation function is applied to the structure with this probability.

keep_non_mutated::Bool: If True, when a mutation happens, the structure before and after the mutation is kept; otherwise only the mutated structure is kept.

MOLGA.Configuration.create_atom_config — Method

create_atom_config(atoms::Dict{String,Int}, molecules::Dict{String,Int}, config_path::String)

Creates AtomConfig from the TOML raw values (atoms, molecules). Elements are internally represented via their atomic number. The molecules are loaded and parsed from their xyz-files.

The config_path is the folder where the config file is stored. The molecule file paths are expected to be relative to this folder.

MOLGA.Configuration.load_config — Method

load_config(configuration_file)

Loads the configuration parameters from the specified TOML file.

I/O

MOLGA.IOm.export_xyz — Method

export_xyz(structures, xyz_file)

Export a set of structures to a xyz files (with element symbols).

Arguments

structures::AbstractVector{Structure}

xyz_file::String: The expected name/path of the file.

MOLGA.IOm.load_xyz — Method

load_xyz(xyz_file::String)

Get all atoms from a structure in XYZ-format (with element symbols).

Helpers

MOLGA.Helpers.atomic_number — Method

atomic_number(element_symbol)

Get the atomic number of an element of the periodic table as Int.

Arguments

element_symbol::AbstractString: The case-sensitive element symbol.

Example

julia> atomic_number("Ar")
18

MOLGA.Helpers.element_symbol — Method

element_symbol(atomic_number)

Get the symbol of an element of the periodic table as String.

Arguments

atomic_number::Int

Example

julia> element_symbol(1)
"H"