API Reference

This page provides a list of all documented functions, structs and modules in the MolecularGeneticAlgorithm. Note that this is list is auto-generated using Documenter's @autodocs function.

MOLGA.Configuration.create_atom_config

MOLGA.Configuration.load_config

MOLGA.GeneticAlgorithm.InitialPopulation.create

MOLGA.GeneticAlgorithm.InitialPopulation.find_position!

MOLGA.GeneticAlgorithm.InitialPopulation.find_position!

MOLGA.GeneticAlgorithm.InitialPopulation.generate_atoms

MOLGA.GeneticAlgorithm.InitialPopulation.generate_indices

MOLGA.GeneticAlgorithm.InitialPopulation.random_angles

MOLGA.GeneticAlgorithm.InitialPopulation.random_position

MOLGA.GeneticAlgorithm.InitialPopulation.rotate_position

MOLGA.GeneticAlgorithm.Utils.check_distance

MOLGA.Helpers.atomic_number

MOLGA.Helpers.element_symbol

MOLGA.IO.load_xyz

MOLGA.run

MOLGA.Configuration.AtomConfig

MOLGA.Configuration.ConfigAtom

MOLGA.Configuration.ConfigMolecule

MOLGA.Configuration.ConfigurationObject

MOLGA.Configuration.DistanceThresholds

MOLGA.Configuration.InitialPopulationConfiguration

MOLGA.Types.Atom

MOLGA.Types.BaseAtom

MOLGA.Types.Structure

MOLGA.Types.Vec

Main Module

MOLGA.run — Function

run(configuration_file [, debug_log::Bool = false])

Start the Molecular Genetic Algorithm.

Genetic Algorithm

MOLGA.GeneticAlgorithm.InitialPopulation.create — Method

create(config)

Initialize the population by creating the desired amount of random structures using the InitialPopulationConfiguration.

Arguments

config::ConfigurationObject

Returns

A vector of structures.

MOLGA.GeneticAlgorithm.InitialPopulation.find_position! — Method

find_position!(atom, atom_list, box_size, distance_thresholds, max_attempts;
    permanent_index, fragment_kind)

Try to find a position for an atom within the periodic boundary condition box by respecting distance thresholds. If a proper position is found, the list of atoms is updated in-place by adding the specified atom with its permanent_index and fragment_kind to the atom_list.

Arguments

atom::ConfigAtom: The atom to be placed into the structure.

atom_list::Vector{Atom}: The list of atoms of the structure.

box_size::AbstractVector

distance_thresholds::DistanceThresholds

max_attempts::Int: Maximum number of attempts to find a proper random position.

permanent_index::Int

fragment_kind::Int

Returns

true if the atom could be placed,

false if no suitable position could be found within the maximum number of attempts.

MOLGA.GeneticAlgorithm.InitialPopulation.find_position! — Method

find_position!(fragment, atom_list, box_size, distance_thresholds, max_attempts;
    permanent_index, fragment_kind)

Place a fragment (ConfigMolecule) into the periodic boundary condition box. Perform distance checks for all atoms of the fragment and assign the specified permanent index and fragment kind to all atoms. This method upates the atom_list in-place.

See find_position!(atom, ...) for more details.

MOLGA.GeneticAlgorithm.InitialPopulation.generate_atoms — Method

generate_atoms(population_config, distance_thresholds, max_iterations)

Distribute the atoms and fragments randomly within a cuboid periodic boundary condition box with respect to the distance thresholds.

Arguments

population_config::InitialPopulationConfiguration

distance_thresholds::DistanceThresholds

max_iterations::Int: The maximum number of iterations to find proper positions while distributing the atoms and fragments.

Returns

A vector of atoms.

MOLGA.GeneticAlgorithm.InitialPopulation.generate_indices — Method

generate_indices(atom_config)

Create a list of the indices using the AtomConfig.

The tuples in the list contain the fragment_kind first and the permanent_index second. Both indices are integers starting from $1$. The entries are inserted as often as specified in the AtomConfig quantity-property.

Each fragment or single atom (one or more) have their own fragment_kind which is equivalent to the index of the AtomConfig list. To provide an option to preserve fragments, all atoms of an instance of a fragment have a common permanent_index.

Example

Small example distributing two hydrogen atoms, one oxygen atom, as well as two carbon dioxide fragments for demonstration purposes.

atom_config::AtomConfig = [
    ConfigAtom(2, 1), # 2x H
    ConfigAtom(1, 8), # 1x O
    ConfigMolecule(
        2, # 2x CO2
        [
            BaseAtom(6, [-1.161937300 0.000000000 -1.704638554]),
            BaseAtom(8, [-0.610097929 -0.126832008 -0.681728275]),
            BaseAtom(8, [-1.713776672 0.126832008 -2.727548833]),
        ],
    ),
]
generate_indices(atom_config)

# output

5-element Vector{Tuple{Int64, Int64}}:
 (1, 1)
 (1, 2)
 (2, 3)
 (3, 4)
 (3, 5)

MOLGA.GeneticAlgorithm.InitialPopulation.random_angles — Method

random_angles()

Generate a vector of three random angles (from $0$ to $2\pi$, in radians).

Example

julia> random_angles()
3-element Vector{Float64}:
 5.892339839566983
 3.042066787260773
 0.8485697032585366

MOLGA.GeneticAlgorithm.InitialPopulation.random_position — Method

random_position(box_size)

Generate a random position within the specified box. The origin is the box's center point, so that a box size of $\mathbf{s}=\left(s_x,s_y,s_z\right)$ leads to a random position $\mathbf{x}$

\[-\frac{1}{2}\,\mathbf{s}\leq\mathbf{x}\leq\frac{1}{2}\,\mathbf{s} \text{.}\]

Example

julia> random_position([8, 5, 5])
3-element StaticArraysCore.SVector{3, Float64} with indices SOneTo(3):
  3.6197648788491597
  1.4205599794740316
 -1.2442172005294312

MOLGA.GeneticAlgorithm.InitialPopulation.rotate_position — Method

rotate_position(position, angle)

Rotate a 3D position vector by a 3D angle vector around the origin using rotation matrices.

Arguments

position::Vector{Float64}: The 3D position vector to be rotated.

angle::Vector{Float64}: The 3D angle vector representing rotation angles around the $x$, $y$, and $z$ axes in radians.

Example

julia> rotate_position([1, 0, 0.5], [0, 0, π])
3-element Vector{Float64}:
 -1.0
  1.2246467991473532e-16
  0.5

MOLGA.GeneticAlgorithm.Utils.check_distance — Method

check_distance(new_position, atoms, thresholds)

Checks if the new position is sufficiently far away from all existing atoms in the atom list.

Interfaces

Types

MOLGA.Types.Atom — Type

Atom(element, position, permanent_index, fragment_kind)

Stores properties of a single atom.

element::Int64: Atomic number of the element, eg. 6 for Carbon.

position::StaticArraysCore.SVector{3, Float64}: 3D position vector of the atom in angstroms.

permanent_index::Int64: Unique index attached to an atom.

fragment_kind::Int64: Kind of fragment this atom belongs to, e.g. all atoms that belong to a H2O molecule have the same value here, also all single O atoms have the same value here, respectively.

MOLGA.Types.BaseAtom — Type

BaseAtom(element, position)

Stores a basic atom.

element::Int64: Atomic number of the element, eg. 6 for Carbon.

position::StaticArraysCore.SVector{3, Float64}: 3D position vector of the atom in angstroms.

MOLGA.Types.Structure — Type

Structure(atoms)
Structure(atoms [, energy [, nuclear_repulsion_energy [, age]]])

Stores properties of a molecular structure.

atoms::Vector{MOLGA.Types.Atom}: Array of atoms that the structure consists of.

energy::Float64: Energy of the structure in eV.

nuclear_repulsion_energy::Float64: Nuclear repulsion energy of the structure in eV.

age::Int64: Number of cycles that this structure is already present.

MOLGA.Types.Vec — Type

Three-dimensional vector.

Configuration

MOLGA.Configuration.AtomConfig — Type

The atom configuration consists of single atoms and molecules.

MOLGA.Configuration.ConfigAtom — Type

ConfigAtom(num, element)

Atom to define initial atom configuration.

quantity::Int64: Number of atoms of this kind.

element::Int64: Chemical element denoted as atomic number.

MOLGA.Configuration.ConfigMolecule — Type

ConfigMolecule(num, atoms)

Molecule to define initial atom configuration.

quantity::Int64: Number of molecules of this kind.

atoms::Vector{MOLGA.Types.BaseAtom}: Array of atoms that the molecule consists of.

MOLGA.Configuration.ConfigurationObject — Type

Stores the entire configuration parameters.

initial_population: See InitialPopulationConfiguration.

charge: Charge of the structure.

spin_multiplicity: Spin multiplicity of the structure.

distance_thresholds: See DistanceThresholds.

MOLGA.Configuration.DistanceThresholds — Type

DistanceThresholds(min, max)

Defines the thresholds for distances between atoms.

min::Float64: Minimum distance that two atoms need.

max::Float64: Maximum distance that two atoms need.

The values are taken into account at several points, including in the generation of the initial population and in mutation and recombination. They are given in angstroms.

MOLGA.Configuration.InitialPopulationConfiguration — Type

Defines the configuration of the initial population.

num_structures::Int64: The number of initial structures to be created.

box_size::StaticArraysCore.SVector{3, Float64}: The three dimensions of the periodic boundary condition box (in angstroms).

atom_config::Array{Union{MOLGA.Configuration.ConfigAtom, MOLGA.Configuration.ConfigMolecule}}: See AtomConfig.

MOLGA.Configuration.create_atom_config — Method

create_atom_config(atoms::Dict{String,Int}, molecules::Dict{String,Int}, config_path::String)

Creates AtomConfig from the TOML raw values (atoms, molecules). Elements are internally represented via their atomic number. The molecules are loaded and parsed from their xyz-files.

The config_path is the folder where the config file is stored. The molecule file paths are expected to be relative to this folder.

MOLGA.Configuration.load_config — Method

load_config(configuration_file)

Loads the configuration parameters from the specified TOML file.

I/O

MOLGA.IO.load_xyz — Method

load_xyz(xyz_file::String)

Get all atoms from a structure in XYZ-format (with element symbols).

Helpers

MOLGA.Helpers.atomic_number — Method

atomic_number(element_symbol)

Get the atomic number of an element of the periodic table as Int.

Arguments

element_symbol::AbstractString: The case-sensitive element symbol.

Example

julia> atomic_number("Ar")
18

MOLGA.Helpers.element_symbol — Method

element_symbol(atomic_number)

Get the symbol of an element of the periodic table as String.

Arguments

atomic_number::Int

Example

julia> element_symbol(1)
"H"